fix miyka-initialize make target

+++ scripts/miyka-link.sh
@@ -5,7 +5,10 @@ TARGET="$2"
 
 if test -e "$TARGET"
 then
-    rm -rf "$SOURCE"
+    if test -L "$TARGET"
+    then
+        rm -rf "$SOURCE"
+    fi
 else
     ln -srf "$SOURCE" "$TARGET"
 fi
